Implemented calculation of productivity of mines (fixes #151) and reworked mine's food consumption.
