+++ jsettlers.common/src/jsettlers/common/buildings/jobs/IBuildingJob.java
@@ -97,4 +97,6 @@ public interface IBuildingJob {
 	String getName();
 
 	boolean isTakeMaterialFromMap();
+
+	EMaterialType[] getFoodOrder();
 }
