+++ jsettlers.common/src/jsettlers/common/buildings/loader/BuildingJobData.java
@@ -95,4 +95,6 @@ public interface BuildingJobData {
 
 	boolean isTakeMaterialFromMap();
 
+	EMaterialType[] getFoodOrder();
+
 }
