+++ jsettlers.common/src/jsettlers/common/buildings/loader/SimpleBuildingJob.java
@@ -27,25 +27,21 @@ import jsettlers.common.movable.EDirection;
 
 public class SimpleBuildingJob implements IBuildingJob {
 	private final short dx;
-
 	private final short dy;
 
 	private final float time;
 
 	private final EMaterialType material;
-
 	private final EDirection direction;
-
 	private final EBuildingJobType type;
-
 	private final ESearchType search;
+	private final String name;
+	private final boolean takeMaterialFromMap;
+
+	private final EMaterialType[] foodOrder;
 
 	private IBuildingJob successJob;
 	private IBuildingJob failJob;
-
-	private final String name;
-
-	private final boolean takeMaterialFromMap;
 
 	/**
 	 * Creates a new simple building job with the given data.
@@ -63,6 +59,7 @@ public class SimpleBuildingJob implements IBuildingJob {
 		search = data.getSearchType();
 		name = data.getName();
 		takeMaterialFromMap = data.isTakeMaterialFromMap();
+		foodOrder = data.getFoodOrder();
 	}
 
 	private SimpleBuildingJob(EBuildingJobType type, int time) {
@@ -75,7 +72,7 @@ public class SimpleBuildingJob implements IBuildingJob {
 		this.search = null;
 		this.name = "";
 		this.takeMaterialFromMap = true;
-
+		this.foodOrder = null;
 	}
 
 	@Override
@@ -209,7 +206,13 @@ public class SimpleBuildingJob implements IBuildingJob {
 		return name;
 	}
 
+	@Override
 	public boolean isTakeMaterialFromMap() {
 		return takeMaterialFromMap;
+	}
+
+	@Override
+	public EMaterialType[] getFoodOrder() {
+		return foodOrder;
 	}
 }
